quant/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
schedule>=1.2.0
yfinance>=0.2.18
supabase>=2.0.0
python-dateutil>=2.8.0
ta-lib>=0.4.25
matplotlib>=3.7.0
seaborn>=0.12.0